+++ database/gormdb.go
@@ -145,574 +145,6 @@ func (db *GormDB) UpdateAccessToken(remote *pb.RemoteIdentity) error {
 	return tx.Commit().Error
 }
 
-// GetUserByRemoteIdentity fetches user by remote identity.
-func (db *GormDB) GetUserByRemoteIdentity(remote *pb.RemoteIdentity) (*pb.User, error) {
-	tx := db.conn.Begin()
-
-	// Get the remote identity.
-	var remoteIdentity pb.RemoteIdentity
-	if err := tx.
-		Where(&pb.RemoteIdentity{
-			Provider: remote.Provider,
-			RemoteID: remote.RemoteID,
-		}).
-		First(&remoteIdentity).Error; err != nil {
-		tx.Rollback()
-		return nil, err
-	}
-
-	// Get the user.
-	var user pb.User
-	if err := tx.Preload("RemoteIdentities").First(&user, remoteIdentity.UserID).Error; err != nil {
-		tx.Rollback()
-		return nil, err
-	}
-
-	if err := tx.Commit().Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-/// Users ///
-
-// GetUser fetches a user by ID with remote identities.
-func (db *GormDB) GetUser(uid uint64) (*pb.User, error) {
-	var user pb.User
-	if err := db.conn.Preload("RemoteIdentities").First(&user, uid).Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-// GetUserWithEnrollments returns user with the given ID with all enrollments.
-func (db *GormDB) GetUserWithEnrollments(uid uint64) (*pb.User, error) {
-	var user pb.User
-	if err := db.conn.Preload("Enrollments").First(&user, uid).Error; err != nil {
-		return nil, err
-	}
-	return &user, nil
-}
-
-// GetUsers fetches all users by provided IDs.
-func (db *GormDB) GetUsers(uids ...uint64) ([]*pb.User, error) {
-	m := db.conn
-	if len(uids) > 0 {
-		m = m.Where(uids)
-	}
-	m = m.Preload("RemoteIdentities")
-
-	var users []*pb.User
-	if err := m.Find(&users).Error; err != nil {
-		return nil, err
-	}
-	return users, nil
-}
-
-// UpdateUser updates user information.
-func (db *GormDB) UpdateUser(user *pb.User) error {
-	if err := db.conn.First(&pb.User{ID: user.GetID()}).Error; err != nil {
-		return err
-	}
-	return db.conn.Save(&user).Error
-}
-
-/// Courses ///
-
-// CreateCourse creates a new course if user with given ID is admin, enrolls user as course teacher.
-func (db *GormDB) CreateCourse(uid uint64, course *pb.Course) error {
-	user, err := db.GetUser(uid)
-	if err != nil {
-		return err
-	}
-	if !user.IsAdmin {
-		return ErrInsufficientAccess
-	}
-
-	var courses uint64
-	if err := db.conn.Model(&pb.Course{}).Where(&pb.Course{
-		OrganizationID: course.OrganizationID,
-	}).Count(&courses).Error; err != nil {
-		return err
-	}
-	if courses > 0 {
-		return ErrCourseExists
-	}
-
-	//TODO(meling) these db updates should be done as a transaction
-	if err := db.conn.Create(course).Error; err != nil {
-		return err
-	}
-	if err := db.CreateEnrollment(&pb.Enrollment{UserID: uid, CourseID: course.ID}); err != nil {
-		return err
-	}
-	if err := db.EnrollTeacher(uid, course.ID); err != nil {
-		return err
-	}
-	return nil
-}
-
-// GetCourse fetches course by ID. If withInfo is true, preloads course
-// assignments, active enrollments and groups.
-func (db *GormDB) GetCourse(cid uint64, withInfo bool) (*pb.Course, error) {
-	m := db.conn
-	var course pb.Course
-
-	if withInfo {
-		// we only want submission from users enrolled in the course
-		userStates := []pb.Enrollment_UserStatus{
-			pb.Enrollment_STUDENT,
-			pb.Enrollment_TEACHER,
-		}
-		// and only group submissions from approved groups
-		modelGroup := &pb.Group{Status: pb.Group_APPROVED, CourseID: cid}
-		if err := m.Preload("Assignments").Preload("Enrollments", "status in (?)", userStates).Preload("Enrollments.User").Preload("Groups", modelGroup).First(&course, cid).Error; err != nil {
-			return nil, err
-		}
-	} else {
-		if err := m.First(&course, cid).Error; err != nil {
-			return nil, err
-		}
-	}
-	db.updateAccessTokenCache(&course)
-	return &course, nil
-}
-
-// GetCourseByOrganizationID fetches course by organization ID.
-func (db *GormDB) GetCourseByOrganizationID(did uint64) (*pb.Course, error) {
-	var course pb.Course
-	if err := db.conn.First(&course, &pb.Course{OrganizationID: did}).Error; err != nil {
-		return nil, err
-	}
-	db.updateAccessTokenCache(&course)
-	return &course, nil
-}
-
-// GetCourses returns a list of courses. If one or more course ids are provided,
-// the corresponding courses are returned. Otherwise, all courses are returned.
-func (db *GormDB) GetCourses(cids ...uint64) ([]*pb.Course, error) {
-	m := db.conn
-	if len(cids) > 0 {
-		m = m.Where(cids)
-	}
-	var courses []*pb.Course
-	if err := m.Find(&courses).Error; err != nil {
-		return nil, err
-	}
-	return courses, nil
-}
-
-// GetCoursesByUser returns all courses (with enrollment status)
-// for the given user id.
-// If enrollment statuses is provided, the set of courses returned
-// is filtered according to these enrollment statuses.
-func (db *GormDB) GetCoursesByUser(uid uint64, statuses ...pb.Enrollment_UserStatus) ([]*pb.Course, error) {
-	enrollments, err := db.getEnrollments(&pb.User{ID: uid}, statuses...)
-	if err != nil {
-		return nil, err
-	}
-
-	var courseIDs []uint64
-	m := make(map[uint64]*pb.Enrollment)
-	for _, enrollment := range enrollments {
-		m[enrollment.CourseID] = enrollment
-		courseIDs = append(courseIDs, enrollment.CourseID)
-	}
-
-	if len(statuses) == 0 {
-		courseIDs = nil
-	} else if len(courseIDs) == 0 {
-		// No need to query database since user have no enrolled courses.
-		return []*pb.Course{}, nil
-	}
-	courses, err := db.GetCourses(courseIDs...)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, course := range courses {
-		course.Enrolled = pb.Enrollment_NONE
-		if enrollment, ok := m[course.ID]; ok {
-			course.Enrolled = enrollment.Status
-		}
-	}
-	return courses, nil
-}
-
-// UpdateCourse updates course information.
-func (db *GormDB) UpdateCourse(course *pb.Course) error {
-	return db.conn.Model(&pb.Course{}).Updates(course).Error
-}
-
-/// Enrollments ///
-
-// CreateEnrollment creates a new pending enrollment.
-func (db *GormDB) CreateEnrollment(enrollment *pb.Enrollment) error {
-	var user, course uint64
-	if err := db.conn.Model(&pb.User{}).Where(&pb.User{
-		ID: enrollment.UserID,
-	}).Count(&user).Error; err != nil {
-		return err
-	}
-	if err := db.conn.Model(&pb.Course{}).Where(&pb.Course{
-		ID: enrollment.CourseID,
-	}).Count(&course).Error; err != nil {
-		return err
-	}
-	if user+course != 2 {
-		return gorm.ErrRecordNotFound
-	}
-
-	enrollment.Status = pb.Enrollment_PENDING
-	return db.conn.Create(&enrollment).Error
-}
-
-// RejectEnrollment removes the user enrollment from the database.
-func (db *GormDB) RejectEnrollment(uid, cid uint64) error {
-	enrol, err := db.GetEnrollmentByCourseAndUser(cid, uid)
-	if err != nil {
-		return err
-	}
-	return db.conn.Delete(enrol).Error
-}
-
-// EnrollStudent enrolls user as course student.
-func (db *GormDB) EnrollStudent(uid, cid uint64) error {
-	return db.setEnrollment(uid, cid, pb.Enrollment_STUDENT)
-}
-
-// EnrollTeacher enrolls user as course teacher.
-func (db *GormDB) EnrollTeacher(uid, cid uint64) error {
-	return db.setEnrollment(uid, cid, pb.Enrollment_TEACHER)
-}
-
-// SetPendingEnrollment sets enrollment status to pending.
-func (db *GormDB) SetPendingEnrollment(uid, cid uint64) error {
-	return db.setEnrollment(uid, cid, pb.Enrollment_PENDING)
-}
-
-// UpdateGroupEnrollment sets GroupID of a student enrollment to 0.
-func (db *GormDB) UpdateGroupEnrollment(uid, cid uint64) error {
-	return db.conn.
-		Model(&pb.Enrollment{}).
-		Where(&pb.Enrollment{CourseID: cid, UserID: uid}).
-		Update("group_id", uint64(0)).Error
-}
-
-// GetEnrollmentByCourseAndUser returns a user enrollment for the given course ID.
-func (db *GormDB) GetEnrollmentByCourseAndUser(cid uint64, uid uint64) (*pb.Enrollment, error) {
-	var enrollment pb.Enrollment
-	m := db.conn.Preload("Course").Preload("User")
-	if err := m.
-		Where(&pb.Enrollment{
-			CourseID: cid,
-			UserID:   uid,
-		}).
-		First(&enrollment).Error; err != nil {
-		return nil, err
-	}
-	return &enrollment, nil
-}
-
-// GetEnrollmentsByCourse fetches all course enrollments with given statuses.
-func (db *GormDB) GetEnrollmentsByCourse(cid uint64, statuses ...pb.Enrollment_UserStatus) ([]*pb.Enrollment, error) {
-	return db.getEnrollments(&pb.Course{ID: cid}, statuses...)
-}
-
-/// Assignments ///
-
-// CreateAssignment creates a new assignment record.
-func (db *GormDB) CreateAssignment(assignment *pb.Assignment) error {
-	// Course id and assignment order must be given.
-	if assignment.CourseID < 1 || assignment.Order < 1 {
-		return gorm.ErrRecordNotFound
-	}
-
-	var course uint64
-	if err := db.conn.Model(&pb.Course{}).Where(&pb.Course{
-		ID: assignment.CourseID,
-	}).Count(&course).Error; err != nil {
-		return err
-	}
-	if course != 1 {
-		return gorm.ErrRecordNotFound
-	}
-
-	return db.conn.
-		Where(pb.Assignment{
-			CourseID: assignment.CourseID,
-			Order:    assignment.Order,
-		}).
-		Assign(pb.Assignment{
-			Name:        assignment.Name,
-			Language:    assignment.Language,
-			Deadline:    assignment.Deadline,
-			AutoApprove: assignment.AutoApprove,
-			ScoreLimit:  assignment.ScoreLimit,
-			IsGroupLab:  assignment.IsGroupLab,
-		}).FirstOrCreate(assignment).Error
-}
-
-// GetAssignment returns assignment with the given ID.
-func (db *GormDB) GetAssignment(query *pb.Assignment) (*pb.Assignment, error) {
-	var assignment pb.Assignment
-	if err := db.conn.Where(query).First(&assignment).Error; err != nil {
-		return nil, err
-	}
-	return &assignment, nil
-}
-
-// GetAssignmentsByCourse fetches all assignments for the given course ID.
-func (db *GormDB) GetAssignmentsByCourse(cid uint64) ([]*pb.Assignment, error) {
-	var course pb.Course
-	if err := db.conn.Preload("Assignments").First(&course, cid).Error; err != nil {
-		return nil, err
-	}
-	return course.Assignments, nil
-}
-
-// UpdateAssignments updates assignment information.
-func (db *GormDB) UpdateAssignments(assignments []*pb.Assignment) error {
-	//TODO(meling) Updating the database may need locking?? Or maybe rewrite as a single query or txn.
-	for _, v := range assignments {
-		// this will create or update an existing assignment
-		if err := db.CreateAssignment(v); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-/// Submissions ///
-
-// CreateSubmission creates a new submission record or updates the most
-// recent submission, as defined by the provided submissionQuery.
-// The submissionQuery must always specify the assignment, and may specify the ID of
-// either an individual student or a group, but not both.
-func (db *GormDB) CreateSubmission(submission *pb.Submission) error {
-	// Primary key must be greater than 0.
-	if submission.AssignmentID < 1 {
-		return gorm.ErrRecordNotFound
-	}
-
-	// Either user or group id must be set, but not both.
-	var m *gorm.DB
-	switch {
-	case submission.UserID > 0 && submission.GroupID > 0:
-		return gorm.ErrRecordNotFound
-	case submission.UserID > 0:
-		m = db.conn.First(&pb.User{ID: submission.UserID})
-	case submission.GroupID > 0:
-		m = db.conn.First(&pb.Group{ID: submission.GroupID})
-	default:
-		return gorm.ErrRecordNotFound
-	}
-
-	// Check that user/group with given ID exists.
-	var group uint64
-	if err := m.Count(&group).Error; err != nil {
-		return err
-	}
-
-	// Checks that the assignment exists.
-	var assignment uint64
-	if err := db.conn.Model(&pb.Assignment{}).Where(&pb.Assignment{
-		ID: submission.AssignmentID,
-	}).Count(&assignment).Error; err != nil {
-		return err
-	}
-
-	if assignment+group != 2 {
-		return gorm.ErrRecordNotFound
-	}
-
-	// Make a new submission struct for the database query to check
-	// whether a submission record for the given lab and user/group
-	// already exists. We cannot reuse the incoming submission
-	// because the query would attempt to match all the test result
-	// fields as well.
-	query := &pb.Submission{
-		AssignmentID: submission.GetAssignmentID(),
-		UserID:       submission.GetUserID(),
-		GroupID:      submission.GetGroupID(),
-	}
-
-	// We want the last record as there can be multiple submissions
-	// for the same student/group and lab in the database.
-	if err := db.conn.Last(query, query).Error; err != nil && err != gorm.ErrRecordNotFound {
-		return err
-	}
-
-	// If a submission for the given assignment and student/group already exists, update it.
-	// Otherwise create a new submission record
-	var labSubmission pb.Submission
-	err := db.conn.Where(query).Assign(submission).FirstOrCreate(&labSubmission).Error
-	submission.ID = labSubmission.GetID()
-	return err
-}
-
-// GetSubmission fetches a submission record.
-func (db *GormDB) GetSubmission(query *pb.Submission) (*pb.Submission, error) {
-	var submission pb.Submission
-	if err := db.conn.Where(query).Last(&submission).Error; err != nil {
-		return nil, err
-	}
-	return &submission, nil
-}
-
-// GetSubmissions returns all submissions for the active assignment for the given course.
-func (db *GormDB) GetSubmissions(courseID uint64, query *pb.Submission) ([]*pb.Submission, error) {
-	var course pb.Course
-	if err := db.conn.Preload("Assignments").First(&course, courseID).Error; err != nil {
-		return nil, err
-	}
-
-	// note that, this creates a query with possibly both user and group;
-	// it will only limit the number of submissions returned if both are supplied.
-	q := &pb.Submission{UserID: query.GetUserID(), GroupID: query.GetGroupID()}
-	var latestSubs []*pb.Submission
-	for _, a := range course.Assignments {
-		q.AssignmentID = a.GetID()
-		temp, err := db.GetSubmission(q)
-		if err != nil {
-			if err == gorm.ErrRecordNotFound {
-				continue
-			}
-			return nil, err
-		}
-		latestSubs = append(latestSubs, temp)
-	}
-	return latestSubs, nil
-}
-
-// GetCourseSubmissions returns all individual lab submissions or group submissions for the course ID
-// depending on the provided groupLabs boolean.
-func (db *GormDB) GetCourseSubmissions(cid uint64, groupLabs bool) ([]pb.Submission, error) {
-	m := db.conn
-
-	// fetch the course entry with all associated assignments and active enrollments
-	course, err := db.GetCourse(cid, true)
-	if err != nil {
-		return nil, err
-	}
-
-	// get IDs of all individual labs or group labs for the course
-	courseAssignmentIDs := make([]uint64, 0)
-	for _, a := range course.Assignments {
-		if a.IsGroupLab == groupLabs {
-			// collect either group labs or non-group labs
-			// but not both in the same collection.
-			courseAssignmentIDs = append(courseAssignmentIDs, a.GetID())
-		}
-	}
-
-	var allLabs []pb.Submission
-	if err := m.Where("assignment_id IN (?)", courseAssignmentIDs).Find(&allLabs).Error; err != nil {
-		return nil, err
-	}
-
-	return allLabs, nil
-}
-
-// UpdateSubmission updates submission with the given approved status.
-func (db *GormDB) UpdateSubmission(sid uint64, approved bool) error {
-	return db.conn.
-		Model(&pb.Submission{}).
-		Where(&pb.Submission{ID: sid}).
-		Update("approved", approved).Error
-}
-
-/// Repositories ///
-
-// CreateRepository creates a new repository record.
-func (db *GormDB) CreateRepository(repo *pb.Repository) error {
-	if repo.OrganizationID == 0 || repo.RepositoryID == 0 {
-		// both organization and repository must be non-zero
-		return ErrCreateRepo
-	}
-
-	switch {
-	case repo.UserID > 0:
-		// check that user exists before creating repo in database
-		err := db.conn.First(&pb.User{}, repo.UserID).Error
-		if err != nil {
-			return err
-		}
-	case repo.GroupID > 0:
-		// check that group exists before creating repo in database
-		err := db.conn.First(&pb.Group{}, repo.GroupID).Error
-		if err != nil {
-			return err
-		}
-	case !repo.RepoType.IsCourseRepo():
-		// both user and group unset, then repository type must be an autograder repo type
-		return ErrCreateRepo
-	}
-
-	return db.conn.Create(repo).Error
-}
-
-// GetRepositoryByRemoteID fetches repository by provider's ID.
-func (db *GormDB) GetRepositoryByRemoteID(rid uint64) (*pb.Repository, error) {
-	var repo pb.Repository
-	if err := db.conn.First(&repo, &pb.Repository{RepositoryID: rid}).Error; err != nil {
-		return nil, err
-	}
-	return &repo, nil
-}
-
-// GetRepositories fetches all repositories satisfying the given query.
-func (db *GormDB) GetRepositories(query *pb.Repository) ([]*pb.Repository, error) {
-	var repos []*pb.Repository
-	if err := db.conn.Find(&repos, query).Error; err != nil {
-		return nil, err
-	}
-	return repos, nil
-}
-
-// DeleteRepositoryByRemoteID deletes repository by provider's ID
-func (db *GormDB) DeleteRepositoryByRemoteID(rid uint64) error {
-	repo, err := db.GetRepositoryByRemoteID(rid)
-	if err != nil {
-		return err
-	}
-	return db.conn.Delete(repo).Error
-}
-
-// Close closes the gorm database.
-func (db *GormDB) Close() error {
-	return db.conn.Close()
-}
-
-/// misc ///
-
-// getEnrollments is generic helper function that return enrollments for either course and user.
-func (db *GormDB) getEnrollments(model interface{}, statuses ...pb.Enrollment_UserStatus) ([]*pb.Enrollment, error) {
-	if len(statuses) == 0 {
-		statuses = []pb.Enrollment_UserStatus{
-			pb.Enrollment_PENDING,
-			pb.Enrollment_STUDENT,
-			pb.Enrollment_TEACHER,
-		}
-	}
-	var enrollments []*pb.Enrollment
-	if err := db.conn.Preload("User").Preload("Course").Model(model).
-		Where("status in (?)", statuses).
-		Association("Enrollments").
-		Find(&enrollments).Error; err != nil {
-		return nil, err
-	}
-	return enrollments, nil
-}
-
-// setEnrollment updates enrollment status.
-func (db *GormDB) setEnrollment(uid, cid uint64, status pb.Enrollment_UserStatus) error {
-	return db.conn.
-		Model(&pb.Enrollment{}).
-		Where(&pb.Enrollment{CourseID: cid, UserID: uid}).
-		Update(&pb.Enrollment{Status: status}).Error
-}
-
 // updateAccessTokenCache caches the access token for the course
 // to allow easy access elsewhere.
 func (db *GormDB) updateAccessTokenCache(course *pb.Course) {
