+++ database/gormdb_assignment.go
@@ -1,0 +1,70 @@
+package database
+
+import (
+	pb "github.com/autograde/aguis/ag"
+	"github.com/jinzhu/gorm"
+)
+
+/// Assignments ///
+
+// CreateAssignment creates a new assignment record.
+func (db *GormDB) CreateAssignment(assignment *pb.Assignment) error {
+	// Course id and assignment order must be given.
+	if assignment.CourseID < 1 || assignment.Order < 1 {
+		return gorm.ErrRecordNotFound
+	}
+
+	var course uint64
+	if err := db.conn.Model(&pb.Course{}).Where(&pb.Course{
+		ID: assignment.CourseID,
+	}).Count(&course).Error; err != nil {
+		return err
+	}
+	if course != 1 {
+		return gorm.ErrRecordNotFound
+	}
+
+	return db.conn.
+		Where(pb.Assignment{
+			CourseID: assignment.CourseID,
+			Order:    assignment.Order,
+		}).
+		Assign(pb.Assignment{
+			Name:        assignment.Name,
+			Language:    assignment.Language,
+			Deadline:    assignment.Deadline,
+			AutoApprove: assignment.AutoApprove,
+			ScoreLimit:  assignment.ScoreLimit,
+			IsGroupLab:  assignment.IsGroupLab,
+		}).FirstOrCreate(assignment).Error
+}
+
+// GetAssignment returns assignment with the given ID.
+func (db *GormDB) GetAssignment(query *pb.Assignment) (*pb.Assignment, error) {
+	var assignment pb.Assignment
+	if err := db.conn.Where(query).First(&assignment).Error; err != nil {
+		return nil, err
+	}
+	return &assignment, nil
+}
+
+// GetAssignmentsByCourse fetches all assignments for the given course ID.
+func (db *GormDB) GetAssignmentsByCourse(cid uint64) ([]*pb.Assignment, error) {
+	var course pb.Course
+	if err := db.conn.Preload("Assignments").First(&course, cid).Error; err != nil {
+		return nil, err
+	}
+	return course.Assignments, nil
+}
+
+// UpdateAssignments updates assignment information.
+func (db *GormDB) UpdateAssignments(assignments []*pb.Assignment) error {
+	//TODO(meling) Updating the database may need locking?? Or maybe rewrite as a single query or txn.
+	for _, v := range assignments {
+		// this will create or update an existing assignment
+		if err := db.CreateAssignment(v); err != nil {
+			return err
+		}
+	}
+	return nil
+}
